Add real-time updates for HA states
Use the home assistant websocket API for listening and handling
state updates. this allows for real-time updates to be received over
websockets rather than polling over the REST API.

For now the only listeners registered are for updating the relevant
zustand state, which is then reflected through the UI in the
grow iot gateway section. Registration happens on useEffect after
initial grows and gateway data is fetched, as well as when entities
are linked. Unlink triggers unregister